requirements.txt
allure-pytest==2.8.19
allure-python-commons==2.8.19
atomicwrites==1.4.0
attrs==20.3.0
importlib-metadata==2.0.0
more-itertools==8.6.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==5.1.3
pytest-html==3.0.0
pytest-metadata==1.10.0
selenium==3.141.0
six==1.15.0
urllib3==1.26.2
wcwidth==0.2.5
zipp==3.4.0
